% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_condition_and_sig_tx.R
\name{add_condition_and_sig_tx}
\alias{add_condition_and_sig_tx}
\title{Add group condition and significance}
\usage{
add_condition_and_sig_tx(alpha = 0.05, info, countsmelted, main_df)
}
\arguments{
\item{alpha}{double to define alpha for significance threshold. Default is 0.05}

\item{info}{dataframe holding sample information. Must have all required column names; therefore must be generated by DTU::create_cample_info. Will be used to join the significance information by sample_id.}

\item{countsmelted}{df from DTU::plot_genes_cohort}

\item{main_df}{dataframe generated by DTU::stageResPlusInfo, holding all results.}
}
\value{
dataframe to be used in DTU::plot_genes_cohort()
}
\description{
Used for generating dataframe for plotting relative transcript abundances in DTU::plot_genes_cohort()
}
