% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_info.R
\name{create_sample_info}
\alias{create_sample_info}
\title{Create phenotype dataframe}
\usage{
create_sample_info(
  pheno,
  cohort_names,
  conditions,
  batch_vars = NULL,
  cohort_col_name = "procedence"
)
}
\arguments{
\item{pheno}{character string representing path to phenotype csv file}

\item{cohort_names}{vector of strings containing cohort names}

\item{conditions}{vector of strings containing names of group conditions (first control then case: c("CT,"CS"))}

\item{batch_vars}{char vector with column names of covariates that should also be read from file}
}
\value{
list of df with minimum 3 columns: sample_id, condition, rin. And additional covariate columns Length of list == length of \code{cohort_names}
}
\description{
TODO: REWRITE THIS!!! simple dplyr and split...Create the sample information, read from the metadatafile
}
