% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitted_bind_observed.R
\name{extract_fitted_bind_observed}
\alias{extract_fitted_bind_observed}
\title{Prepare df to plot fitted proportions}
\usage{
extract_fitted_bind_observed(ds_obj, dge_obj = NULL, observed_counts, tool)
}
\arguments{
\item{ds_obj}{result object from either DRIMSeq or DEXSeq}

\item{dge_obj}{DESeq2 result object}

\item{observed_counts}{dataframe of counts unfitted  (from DTU::extract_raw_counts())}

\item{tool}{<"drim"|"dex"> specifying the tool from which to extract the fitted values}
}
\description{
decide which tool to extract fitted values from
extract them
if there is a DESeq2 obj (dge), extract fitted values from there aswell
bind fitted, (dge) and observed
}
