% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raw_counts.R
\name{extract_raw_counts}
\alias{extract_raw_counts}
\title{Extract counts from tool object}
\usage{
extract_raw_counts(obj, tool, gene_info)
}
\arguments{
\item{obj}{either DRIMSeq or DEXSeq object}

\item{tool}{<"drim"|"dex">}

\item{gene_info}{df with gene annotation from DTU::get_gene_info()}
}
\value{
dataframe with raw counts
}
\description{
Used for plotting. Extract counts from respective tools
}
