% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sig.R
\name{extract_sig}
\alias{extract_sig}
\title{extract significant hits}
\usage{
extract_sig(obj, alpha = 0.05)
}
\arguments{
\item{obj}{list of stageR result dataframes, where transcript is the transcript p-value adjusted by stageR and gene is the column for the gene-level adjusted pvalue}
}
\value{
list of dataframes holding only the significant events
}
\description{
Extract significant events
}
\details{
Go through list of stageR result dataframes to extract significant events. Not sure where I use this, since stageR has a function for this and results from stageR are only valid with the FDR that was used to run stageR
}
