% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annot.R
\name{get_annot}
\alias{get_annot}
\title{create annotation df}
\usage{
get_annot(
  version = "75",
  g_source = "ucsc",
  gtf_path = NULL,
  ntx_filter = NULL
)
}
\arguments{
\item{version}{string ensembldb version (just the number). Default=75}

\item{ntx_filter}{integer threshold for maximum number of transcripts per gene. Genes with more transcripts will be removed from the annotation}

\item{g_scource}{character string genome annotation source, ensembldb or ucsc (->gtf file). Remember there were some modifications needed to be done for the gtf file.}
}
\value{
annotation dataframe as required by DRIMSeq in DTU::prep_dtu() to build DRIMSeqData object
}
\description{
Create transcript annotation file to map transcripts to genes. Either with ensembldb or read from supplied gtf path. Used gtf file in referenceData dir.
}
