% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_counts.R
\name{import_counts}
\alias{import_counts}
\title{tximport wrapper}
\usage{
import_counts(
  salmon_data,
  info,
  quant_tool = "salmon",
  scale_method = "scaledTPM",
  annot
)
}
\arguments{
\item{salmon_data}{base directory of folders with salmon output (one folder per sample), e.g. \"myHome/salmonOutput/\"}

\item{info}{dataframe or list of dataframes holding sample meta data created from DTU::create_sample_info(). One dataframe per cohort.}

\item{quant_tool}{character string indicating which tarnscript quantification tool was used. This is redundant. Always used salmon. Default: salmon}

\item{scale_method}{character string indicating which scaling method should be used by tximport . Default scaledTPM as suggested for DTU analysis.}

\item{annot}{annotation df holding annotations for the transcripts. Created by DTU::get_annot() using a reference file (used the .gtf file in the referenceData dir)}
}
\value{
list of tximport objects, one for each cohort.
}
\description{
Import transcript-level counts from Salmon output directories using sampe IDs from the info dataframe from DTU::create_sample_info
}
