% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_annot.R
\name{plot_gene_annot}
\alias{plot_gene_annot}
\title{plot transcript body}
\usage{
plot_gene_annot(gene_id, tx_ids, gene_name, biotype = NULL)
}
\arguments{
\item{gene_id}{ensembl gene id of the gene to be plotted}

\item{tx_ids}{string vector of transcript ids (ensembl) specifying which of the transcripts of the gene should be plotted}

\item{gene_name}{string gene name for the plot tilte}

\item{biotype}{string specifying transcript biotypes as annotation in the plot. Default is NULL.}
}
\value{
GenomeGraphs plot object
}
\description{
Using biomaRt and GenomeGraphs to plot transcript body
}
