% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_cohort.R
\name{plot_genes_cohort}
\alias{plot_genes_cohort}
\title{Extract counts and plot DTU}
\usage{
plot_genes_cohort(
  out = "./",
  obj,
  want_jitter = TRUE,
  conditions = c("Control", "Case"),
  cohort = NULL,
  selected_samples = NULL,
  genes_toplot = NULL,
  gene_name = NULL,
  tool = "dex",
  plot = T,
  only_nom_sig = T
)
}
\arguments{
\item{out}{string specifying output directory for pdf of plot}

\item{obj}{result data object (see makeFigures.Rmd)}

\item{want_jitter}{bool whether to plot fitted and observed as points jittered (old, not used anymore)}

\item{selected_samples}{string vector of sample ids, of which to extract the counts for plotting}

\item{genes_toplot}{df ; if is.null(gene_name) provide the df from DTU::get_gene_info with all transcripts that should be plotted.
(Df can have rows with transcript from more than one gene, then a list of plots, one for each gene will be returned)}

\item{gene_name}{string specifying the gene which should be plotted}

\item{tool}{<"drim"|"dex">}

\item{plot}{bool, if TRUE, print plot}

\item{only_nom_sig}{bool whether to mark sepcific samples in the plot with a txt label}
}
\description{
extracts fitted and observed counts and plots relative transcript expressions for specific tool and specific cohort
}
