% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveAAs.R
\name{retrieveAAs}
\alias{retrieveAAs}
\title{Retrieve protein sequence}
\usage{
retrieveAAs(tx_ids, eV = "75", out = "./test.fasta", append = F)
}
\arguments{
\item{tx_ids}{string vector with transcript ids identifying transcripts for which protein sequence should be extracted. Must be compatible with ensembldb release version specified in eV.}

\item{eV}{string specifying ensembldb release version to be used}

\item{out}{string specifying file to which output should be written.}

\item{append}{bool specifying whether protein sequences should be appended to content in file or if file should be overwritten. Default is FALSE.}
}
\value{
dataframe with protein sequences
}
\description{
Get protein sequence of transcript isoforms from ensembldb
}
