% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dex.R
\name{run_dex}
\alias{run_dex}
\title{Wrapper for DEXSeq}
\usage{
run_dex(info, obj, covariates = list("rin"))
}
\arguments{
\item{info}{<df> or list of df with 3 columns (sample_id,condition,rin) and additional covariates}

\item{obj}{list of DRIMSeqData objects created by DTU::prep_dtu(). One for each cohort.}

\item{covariates}{<list> names of variables to include into model design (actual values must be present in info)}
}
\value{
list of DEXSeq objects after DTU analysis
TODO: if covariates is null
}
\description{
Used in Rscript scripts/"dexStage.R". Builds model matrix from sample info and runs DEXSeq. Countdata is ecxtracted from DRIMSeq object from DTU::prep_dtu() after filtering.
}
