% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_drim.R
\name{run_drim}
\alias{run_drim}
\title{Wrapper for DRIMSeq}
\usage{
run_drim(info, obj, covariates)
}
\arguments{
\item{info}{<df> or list of df with 3 columns (sample_id,condition,rin) and additional covariates}

\item{obj}{list of DRIMSeqData objects created by DTU::prep_dtu(). One for each cohort.}

\item{covariates}{<list> names of variables to include into model design (actual values must be present in info)}
}
\value{
list of DRIMSeq objects after DTU analysis
TODO: if covariates is null
}
\description{
Used in Rscript scripts/"drimStage.R". Builds model matrix from sample info and runs DRIMSeq. Countdata is ecxtracted from DRIMSeq object from DTU::prep_dtu() after filtering.
}
