% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_intersect.R
\name{simple_intersect}
\alias{simple_intersect}
\title{Intersect}
\usage{
simple_intersect(res_obj, tx = TRUE, gene_col = "geneID", tx_col = "txID")
}
\arguments{
\item{res_obj}{list of dataframes resulting from stageR}

\item{tx}{bool, if TRUE, intersects transcript ids, else gene ids}

\item{gene_col}{character string of gene id column name}

\item{tx_col}{character string of transcrip id column name}
}
\value{
character string holding strings which overlap
}
\description{
Determine intersection of elements in vectors in a list
}
