% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_cohorts.R
\name{split_by_cohorts}
\alias{split_by_cohorts}
\title{Reorder nested list with DTU objects}
\usage{
split_by_cohorts(cohort_names, objs)
}
\arguments{
\item{cohort_names}{char vector with names of cohorts (as used in the pipeline (bash))}

\item{objs}{the nested list that should be reorderd}
}
\value{
nested list with the outer list named by cohorts, and a list with specific tool results in each of those
}
\description{
Used in the Rscript "bindToolRes.R". Transfrom listing order from 'per tool' to 'per cohort'.
Input obj consists of a list for each tool, and the tool specific list holds gene-lists for each cohort.
The function swapps this around such that the outer of the nested list is a list of cohorts within which the two tool specific gene results are listed
}
