% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_colon.R
\name{split_colon}
\alias{split_colon}
\title{split string by colon 
Really unnecessary function that probably already exists. Used to split the input argument strings from the commandline in runDTU.sh. Arguments like for example cohort_names are specified in the bash script with "discovery:replication".}
\usage{
split_colon(string)
}
\arguments{
\item{character}{string that should be split by colon}
}
\value{
named character vector
}
\description{
split string by colon 
Really unnecessary function that probably already exists. Used to split the input argument strings from the commandline in runDTU.sh. Arguments like for example cohort_names are specified in the bash script with "discovery:replication".
}
\examples{
split_colon("Control:Case")
split_colon("gender:age:rin")
}
