% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageResPlusInfo.R
\name{stageResPlusInfo}
\alias{stageResPlusInfo}
\title{Generate main result dataframe}
\usage{
stageResPlusInfo(genelists, ds, cohort)
}
\arguments{
\item{cohort}{string cohort name to be used to extract from the lists}

\item{list}{of DRIMSeq and DEXSeq objects}
}
\value{
list of dataframes (one for each tool) holding all relevant result information 
(nominal p-values, stageR screened p-values, (one for each tool) condition coefficient (adjusted for DRIMSeq (log2(exp(coefficient)))))
}
\description{
Makes a df to more easily compare tool and cohort results. Df have same colnames so can be melted etc for plotting
 requires stageR result lists, the drimseqdexseq objects and which cohort to use
}
